M3-1894 renamed interaction params to email params
